+++ src/file_path.rs
@@ -92,7 +92,6 @@ impl From<std::path::PathBuf> for FilePath {
     }
 }
 
-
 impl Into<std::path::PathBuf> for FilePath {
     fn into(self) -> std::path::PathBuf {
         std::path::PathBuf::from(self.as_os_str())